@@ -1,6 +1,6 @@
 package com.yjh.rer.base;
 
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 
 import javax.inject.Inject;
 
