@@ -1,23 +1,22 @@
 package com.yjh.rer.base;
 
-import android.arch.lifecycle.LifecycleRegistry;
-import android.arch.lifecycle.LifecycleRegistryOwner;
+import androidx.lifecycle.LifecycleOwner;
+import androidx.lifecycle.LifecycleRegistry;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentTransaction;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentTransaction;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 //import com.squareup.leakcanary.RefWatcher;
-import com.yjh.rer.MyApplication;
 import com.yjh.rer.R;
 
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
 
-public abstract class BaseFragment extends Fragment implements LifecycleRegistryOwner {
+public abstract class BaseFragment extends Fragment implements LifecycleOwner {
 
     private LifecycleRegistry mLifecycleRegistry = new LifecycleRegistry(this);
     private Unbinder mUnBinder;
