@@ -1,8 +1,8 @@
 package com.yjh.rer.network;
 
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 public class Resource<T> {
     private final Status mStatus;
