@@ -1,7 +1,7 @@
 package com.yjh.rer.room.db;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
 
 import com.yjh.rer.room.dao.RedEnvelopeDao;
 import com.yjh.rer.room.entity.RedEnvelope;
